Changing transport explicitly to websocket
For some reason socket.io is not upgrading the transport from polling to websocket as it is supposed to. This is causing the load balancer to route the client every time the client polls... Instead, we are starting witha websocket connection and skipping the http upgrade. Need to benchmark after making this change.